content-onboarding/pyproject.toml
[tool.poetry]
name = "biosearch-core"
version = "0.1.0"
description = ""
authors = ["Juan Trelles <juantrelles@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.3"
pyarrow = "^10.0.1"
Pillow = "^9.4.0"
psycopg = {extras = ["binary"], version = "^3.1.8"}
python-dotenv = "^0.21.1"
image_modalities_classifier = { git = "https://github.com/uic-evl/bio-search.git", subdirectory="image-modalities-classifier" }

# optional due to its dependency on CUDA (11.6)
torch = { version="^1.13.1", source="torch", optional = true }
torchvision = { version = "^0.14.1", source="torch", optional = true }
tqdm = "^4.64.1"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu116"
default = false
secondary = true

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.20.2"
black = "^22.12.0"
pylint = "^2.15.10"

[tool.poetry.extras]
torch = ["torch", "torchvision"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
biosearch_core_export = 'biosearch_core.indexing.export:main'